+++ packages/sdk/src/util.ts
@@ -235,3 +235,15 @@ export function isDefined<T>(v?: T): v is T {
 export function toBuffer(s: string, encoding?: string): Buffer {
     return Buffer.from(s, encoding);
 }
+
+/**
+ * Immediately returns an {@linkcode Error} with the message passed. This allows use of throwing errors as expressions.
+ * @param message - The message to pass to the error
+ * @throws an error immediately
+ *
+ * @example
+ * const value = maybeValue ?? bail('Turns out there was not value anyway...');
+ */
+export const bail = (message: string): never => {
+    throw new Error(message);
+};