+++ packages/sdk/src/wasm/web3Id.ts
@@ -1,8 +1,13 @@
 import * as wasm from '@concordium/rust-bindings/wallet';
 import { stringify } from 'json-bigint';
-import { VerifyWeb3IdCredentialSignatureInput } from '../web3-id/web3IdHelpers.js';
-import { Web3IdProofInput } from '../web3-id/web3IdProofTypes.js';
+import { VerifyWeb3IdCredentialSignatureInput } from '../web3-id/helpers.js';
+import {
+    CredentialsInputs,
+    Web3IdProofInput,
+    Web3IdProofRequest,
+} from '../web3-id/types.js';
 import { VerifiablePresentation } from '../types/VerifiablePresentation.js';
+import { CryptographicParameters } from '../types.js';
 
 /**
  * Verifies that the given signature is correct for the given values/randomness/holder/issuerPublicKey/issuerContract
@@ -29,4 +34,24 @@ export function getVerifiablePresentation(
     } catch (e) {
         throw new Error(e as string);
     }
+}
+
+/**
+ * Verify the proofs contained in the {@linkcode VerifiablePresentation}.
+ *
+ * @param presentation - The verifiable presentation to verify.
+ * @param globalContext - The cryptographic parameters of the chain the presentation is created on.
+ * @param publicData - The public data corresponding to the proofs contained in the presentation.
+ *
+ * @returns The request corresponding to the verifable presentation, if verification is successful.
+ * @throws If the verification is not successful for any reason.
+ */
+export function verifyPresentation(
+    presentation: VerifiablePresentation,
+    globalContext: CryptographicParameters,
+    publicData: CredentialsInputs[]
+): Web3IdProofRequest {
+    const input = stringify({ presentation, globalContext, publicData });
+    const result = wasm.verifyPresentation(input);
+    return JSON.parse(result);
 }