+++ packages/sdk/src/web3-id/types.ts
@@ -1,3 +1,4 @@
+import { CIS4 } from '../cis4/util.js';
 import {
     GenericAtomicStatement,
     GenericRevealStatement,
@@ -5,7 +6,11 @@ import {
     GenericNonMembershipStatement,
     GenericRangeStatement,
 } from '../commonProofTypes.js';
-import type { CryptographicParameters } from '../types.js';
+import type {
+    AttributeKey,
+    CryptographicParameters,
+    HexString,
+} from '../types.js';
 import type * as ContractAddress from '../types/ContractAddress.js';
 
 export type TimestampAttribute = {
@@ -247,4 +252,33 @@ export type CredentialStatements = CredentialStatement[];
 export type CredentialSubject = {
     id: string;
     attributes: Record<string, AttributeType>;
+};
+
+/** The credentials inputs required to verify the proof of account proofs */
+export type CredentialsInputsAccount = {
+    /** Union tag */
+    type: 'account';
+    /** Commitments for the ID attributes of the account */
+    commitments: Partial<Record<AttributeKey, HexString>>;
+};
+
+/** The credentials inputs required to verify the proof of Web3 ID proofs */
+export type CredentialsInputsWeb3 = {
+    /** Union tag */
+    type: 'web3';
+    /** The public key of the Web3 ID issuer */
+    issuerPk: HexString;
+};
+
+/** Union of the different inputs required to verify corresponding proofs */
+export type CredentialsInputs =
+    | CredentialsInputsAccount
+    | CredentialsInputsWeb3;
+
+/** Contains the credential status and inputs required to verify a corresponding credential proof */
+export type CredentialWithMetadata = {
+    /** The credential status */
+    status: CIS4.CredentialStatus;
+    /** The public data required to verify a corresponding credential proof */
+    inputs: CredentialsInputs;
 };