+++ packages/rust-bindings/CHANGELOG.md
@@ -4,7 +4,7 @@
 
 ### Added
 
-- The function `verify_presentation` to `@concordium/rust-bindings/wallet`, exposed with the WASM entrpoint `verifyPresentation`.
+- The function `verify_presentation` to `@concordium/rust-bindings/wallet`, exposed with the WASM entrypoint `verifyPresentation`.
 
 ## 3.0.0